+++ packages/flowr/CHANGELOG.md
@@ -17,6 +17,7 @@ refactor hide FrViewModel:: debugDescribeChildren, ::toDiagnosticsNode
 ## 2.0.0 2025-8-29
 * retract v1.4.0 , re publish as v2.0.0
 - ::updateRaw -> ::update is **break** change
+- rename FrProvider.container -> FrProvider.di
 
 ## 1.4.0 2025-8-28
 * feat ::update 'update' can not return null;